+++ 12/bridges.in
@@ -1,0 +1,10 @@
+7 9
+1 2
+2 3
+3 4
+1 3
+4 5
+4 6
+5 6
+6 6
+3 4
