+++ 12/task.sln
@@ -1,0 +1,43 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.489
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "task", "task.vcxproj", "{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Out|x64 = Out|x64
+		Out|x86 = Out|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+		Test|x64 = Test|x64
+		Test|x86 = Test|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Debug|x64.ActiveCfg = Debug|x64
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Debug|x64.Build.0 = Debug|x64
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Debug|x86.ActiveCfg = Debug|Win32
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Debug|x86.Build.0 = Debug|Win32
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Out|x64.ActiveCfg = Out|x64
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Out|x64.Build.0 = Out|x64
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Out|x86.ActiveCfg = Out|Win32
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Out|x86.Build.0 = Out|Win32
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Release|x64.ActiveCfg = Release|x64
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Release|x64.Build.0 = Release|x64
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Release|x86.ActiveCfg = Release|Win32
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Release|x86.Build.0 = Release|Win32
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Test|x64.ActiveCfg = Test|x64
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Test|x64.Build.0 = Test|x64
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Test|x86.ActiveCfg = Test|Win32
+		{B632B834-C6E4-4BF1-8FAA-6E09F8FB2F0F}.Test|x86.Build.0 = Test|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {21D1ACF9-B56A-491E-8241-2252ABEBE740}
+	EndGlobalSection
+EndGlobal
