+++ 12/test.cpp
@@ -1,0 +1,12 @@
+#include <random>
+#include <cassert>
+#include <iostream>
+#include <vector>
+#include "memcheck_crt.h"
+
+
+int main() {
+    ENABLE_CRT;
+
+    std::cout << "tests not implemented" << std::endl;
+}
